Use browser-based checks for updates
